Move `reelmagic` to `src/hardware/video`
