+++ src/dosbox.cpp
@@ -33,8 +33,8 @@
 #include "hardware/pci_bus.h"
 #include "hardware/pic.h"
 #include "hardware/port.h"
-#include "hardware/reelmagic/reelmagic.h"
 #include "hardware/timer.h"
+#include "hardware/video/reelmagic/reelmagic.h"
 #include "hardware/video/video.h"
 #include "hardware/video/voodoo.h"
 #include "ints/int10.h"
