+++ src/hardware/video/reelmagic/tools/is_magical_asset.c
