+++ src/hardware/video/reelmagic/tools/wincompat.h
