+++ src/hardware/video/vga_dac.cpp
@@ -11,7 +11,7 @@
 #include "gui/render.h"
 #include "hardware/memory.h"
 #include "hardware/port.h"
-#include "hardware/reelmagic/reelmagic.h"
+#include "hardware/video/reelmagic/reelmagic.h"
 #include "ints/int10.h"
 #include "utils/rgb.h"
